Implement an authorization web service which acts as a centralized
authorization service for measurements and messages.

+++ Core/Models/DataPoint.cs
@@ -22,7 +22,7 @@ namespace SensateService.Models
 		[BsonRequired]
 		public decimal Value { get; set; }
 		[BsonIgnoreIfNull]
-		public decimal? Precision { get; set; }
+		public double? Precision { get; set; }
 		[BsonIgnoreIfNull]
 		public double? Accuracy { get; set; }
